Move MVT files to Resources (styles, sprites)
