+++ map/public/mvt/sprites
@@ -1,0 +1,1 @@
+../../../../resources/mvt/sprites
