+++ map/src/map/layers/MvtLayerConfig.js
@@ -1,6 +1,6 @@
 export const MVT_OSMAND_LAYER = 'mvt-osmand';
 
-// URL in defaultStyle.json is used for external style editors only
+// URL in default.mvt.json is used for external style editors only
 // This URL overrides the default one and requires setupProxy.js to proxy /vector/ queries
 const port = window.location.port;
 const host = window.location.hostname.toLowerCase();
@@ -10,7 +10,7 @@ const tilesHost = host === 'osmand.net' || host === 'www.osmand.net' ? 'maptile.
 export const MVT_OSMAND_URL = `${window.location.protocol}//${tilesHost}${tilesPort}/vector/{z}/{x}/{y}.mvt${cache}`;
 
 export const MVT_OSM_TEST_LAYER = 'mvt-osm-test';
-export const MVT_OSM_TEST_URL = 'https://tiles.openstreetmap.us/vector/openmaptiles/{z}/{x}/{y}.mvt'; // mvtOsmStyle.json
+export const MVT_OSM_TEST_URL = 'https://tiles.openstreetmap.us/vector/openmaptiles/{z}/{x}/{y}.mvt';
 
 export const mvtOsmAndURL = {
     uiname: 'OsmAnd vector (beta)',
