+++ map/src/resources/mvt
@@ -1,0 +1,1 @@
+../../../../resources/mvt
